+++ tests/Containers/Inventory/InventoryTests.AddItem.cs
@@ -1,7 +1,7 @@
 ﻿using TheChest.Inventories.Slots.Interfaces;
+using TheChest.Tests.Common.Attributes;
 using TheChest.Tests.Common.Extensions.Containers;
 using TheChest.Tests.Common.Extensions.Slots;
-using TheChest.Tests.Common.Attributes;
 
 namespace TheChest.Inventories.Tests.Containers.Inventory
 {
@@ -17,6 +17,25 @@ namespace TheChest.Inventories.Tests.Containers.Inventory
                 () => inventory.Add(item: default!),
                 Throws.ArgumentNullException.With.Property("ParamName").EqualTo("item")
             );
+        }
+
+        [Test]
+        [IgnoreIfReferenceType]
+        public void AddItem_DefaultValue_AddsItemToFirstSlot()
+        {
+            var size = this.GenerateRandomSize();
+            var inventory = this.inventoryFactory.EmptyContainer(size);
+
+            var item = default(T);
+            inventory.Add(item);
+
+            Assert.Multiple(() =>
+            {
+                var firstSlot = inventory.GetSlot(0);
+                Assert.That(firstSlot, Is.Not.Null);
+                Assert.That(firstSlot.IsEmpty, Is.False);
+                Assert.That(firstSlot.GetContent(), Is.EqualTo(item));
+            });
         }
 
         [Test]
@@ -41,14 +60,17 @@ namespace TheChest.Inventories.Tests.Containers.Inventory
             var inventory = this.inventoryFactory.FullContainer(size, items);
 
             var item = this.itemFactory.CreateRandom();
-            Assert.That(() => inventory.Add(item), Throws.InvalidOperationException);
 
-            Assert.That(
-                inventory.GetSlots(),
-                Is.All.Matches<IInventorySlot<T>>(
-                    x => x.IsFull && !x.GetContent()!.Equals(item)
-                )
-            );
+            Assert.Multiple(() =>
+            {
+                Assert.That(() => inventory.Add(item), Throws.InvalidOperationException);
+                Assert.That(
+                    inventory.GetSlots(),
+                    Is.All.Matches<IInventorySlot<T>>(
+                        x => x.IsFull && !x.GetContent()!.Equals(item)
+                    )
+                );
+            });
         }
 
         [Test]