+++ tests/Containers/Inventory/InventoryTests.CanReplace.cs
@@ -5,6 +5,7 @@ namespace TheChest.Inventories.Tests.Containers.Inventory
     public partial class InventoryTests<T>
     {
         [Test]
+        [IgnoreIfValueType]
         public void CanReplace_NullItem_ThrowsArgumentNullException()
         {
             var size = this.GenerateRandomSize();
@@ -14,6 +15,19 @@ namespace TheChest.Inventories.Tests.Containers.Inventory
                 () => inventory.CanReplace(default!, 0),
                 Throws.ArgumentNullException.With.Property("ParamName").EqualTo("item")
             );
+        }
+
+        [Test]
+        [IgnoreIfReferenceType]
+        public void CanReplace_DefaultItem_ReturnsTrue()
+        {
+            var size = this.GenerateRandomSize();
+            var inventory = this.inventoryFactory.FullContainer(size, this.itemFactory.CreateDefault());
+
+            var randomIndex = this.random.Next(0, size);
+            var canReplace = inventory.CanReplace(default!, randomIndex);
+
+            Assert.That(canReplace, Is.True);
         }
 
         [TestCase(-1)]
@@ -42,18 +56,6 @@ namespace TheChest.Inventories.Tests.Containers.Inventory
         }
 
         [Test]
-        public void CanReplace_EmptySelectedSlot_ReturnsTrue()
-        {
-            var size = this.GenerateRandomSize();
-            var inventory = this.inventoryFactory.EmptyContainer(size);
-            var randomIndex = this.random.Next(0, size);
-
-            var canReplace = inventory.CanReplace(this.itemFactory.CreateDefault(), randomIndex);
-
-            Assert.That(canReplace, Is.True);
-        }
-
-        [Test]
         public void CanReplace_FullSelectedSlot_ReturnsTrue()
         {
             var size = this.GenerateRandomSize();
@@ -78,7 +80,7 @@ namespace TheChest.Inventories.Tests.Containers.Inventory
         }
 
         [Test]
-        public void CanReplace_EmptySlot_ReturnsTrue()
+        public void CanReplace_EmptySlot_ReturnsFalse()
         {
             var size = this.GenerateRandomSize();
             var inventory = this.inventoryFactory.EmptyContainer(size);
@@ -86,7 +88,7 @@ namespace TheChest.Inventories.Tests.Containers.Inventory
 
             var result = inventory.CanReplace(this.itemFactory.CreateDefault(), randomIndex);
 
-            Assert.That(result, Is.True);
+            Assert.That(result, Is.False);
         }
     }
 }