+++ tests/Slots/InventorySlot/InventorySlotTests.replace.cs
@@ -21,54 +21,27 @@ namespace TheChest.Inventories.Tests.Slots.InventorySlot
 
         #region Empty Slot
         [Test]
-        [IgnoreIfValueType]
-        public void Replace_EmptySlot_ReturnsNull()
+        public void Replace_EmptySlot_ThrowsInvalidOperationException()
         {
             var slot = this.slotFactory.Empty();
-            var newItem = this.itemFactory.CreateDefault();
 
-            var result = slot.Replace(newItem);
-
-            Assert.That(result, Is.Null);
+            var newItem = this.itemFactory.CreateRandom();
+            Assert.That(
+                () => slot.Replace(newItem),
+                Throws.InvalidOperationException.With.Message.EqualTo("The slot is empty.")
+            );
         }
 
         [Test]
-        [IgnoreIfValueType]
-        public void Replace_EmptySlot_AddsItem()
+        public void Replace_EmptySlot_DoesNotReplacesItem()
         {
             var slot = this.slotFactory.Empty();
-            var newItem = this.itemFactory.CreateDefault();
-
-            slot.Replace(newItem);
-
-            Assert.That(slot.GetContent(), Is.EqualTo(newItem));
-        }
-
-        [Test]
-        [IgnoreIfReferenceType]
-        public void Replace_EmptySlot_DefaultItem_ReturnsDefault()
-        {
-            var slot = this.slotFactory.Empty();
-
-            var newItem = default(T);
-            var result = slot.Replace(newItem!);
-
-            Assert.That(result, Is.EqualTo(default(T)));
-        }
-
-        [Test]
-        [IgnoreIfReferenceType]
-        public void Replace_EmptySlot_DefaultItem_AddsDefault()
-        {
-            var slot = this.slotFactory.Empty();
-
-            var newItem = default(T);
-            slot.Replace(newItem!);
 
             Assert.Multiple(() =>
             {
-                Assert.That(slot.GetContent(), Is.EqualTo(default(T)));
-                Assert.That(slot.IsFull, Is.True);
+                var newItem = this.itemFactory.CreateRandom();
+                Assert.That(() => slot.Replace(newItem), Throws.InvalidOperationException);
+                Assert.That(slot.GetContent(), Is.Not.EqualTo(newItem));
             });
         }
         #endregion
@@ -91,8 +64,8 @@ namespace TheChest.Inventories.Tests.Slots.InventorySlot
         {
             var initialItem = this.itemFactory.CreateDefault();
             var slot = this.slotFactory.Full(initialItem);
-            var newItem = this.itemFactory.CreateRandom();
 
+            var newItem = this.itemFactory.CreateRandom();
             slot.Replace(newItem);
 
             Assert.That(slot.GetContent(), Is.EqualTo(newItem));
@@ -100,7 +73,7 @@ namespace TheChest.Inventories.Tests.Slots.InventorySlot
 
         [Test]
         [IgnoreIfReferenceType]
-        public void Replace_DefaultItem_FullSlot_ReturnsItem()
+        public void Replace_DefaultValue_FullSlot_ReturnsItem()
         {
             var item = this.itemFactory.CreateRandom();
             var slot = this.slotFactory.Full(item);
@@ -113,7 +86,7 @@ namespace TheChest.Inventories.Tests.Slots.InventorySlot
 
         [Test]
         [IgnoreIfReferenceType]
-        public void Replace_DefaultItem_FullSlot_AddsDefault()
+        public void Replace_DefaultValue_FullSlot_AddsDefault()
         {
             var item = this.itemFactory.CreateRandom();
             var slot = this.slotFactory.Full(item);
@@ -130,7 +103,7 @@ namespace TheChest.Inventories.Tests.Slots.InventorySlot
 
         [Test]
         [IgnoreIfReferenceType]
-        public void Replace_DefaultItem_FullSlot_DefaultContent_ReturnsDefault()
+        public void Replace_FullSlotWithDefaultContent_ReturnsDefault()
         {
             var slot = this.slotFactory.Full(default!);
 
@@ -142,7 +115,7 @@ namespace TheChest.Inventories.Tests.Slots.InventorySlot
 
         [Test]
         [IgnoreIfReferenceType]
-        public void Replace_DefaultItem_FullSlot_DefaultContent_AddsItem()
+        public void Replace_FullSlotWithDefaultContent_AddsItem()
         {
             var slot = this.slotFactory.Full(default!);